+++ projects/np-ui-lib/src/lib/np-date-picker/np-date-picker.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NpDatePickerComponent } from './np-date-picker.component';
-import { FormsModule } from '@angular/forms';
 import { NpTooltipModule } from '../np-tooltip/np-tooltip.module';
 import { OverlayModule } from '@angular/cdk/overlay';
 import { PortalModule } from '@angular/cdk/portal';
@@ -11,7 +10,6 @@ import { NpUtilityModule } from '../np-utility/np-utility.module';
   declarations: [NpDatePickerComponent],
   imports: [
     CommonModule,
-    FormsModule,
     OverlayModule,
     PortalModule,
     NpTooltipModule,