+++ projects/np-ui-lib/src/lib/np-file-upload/np-file-upload.module.ts
@@ -1,13 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NpFileUploadComponent } from './np-file-upload.component';
-import { FormsModule } from '@angular/forms';
 
 @NgModule({
   declarations: [NpFileUploadComponent],
   imports: [
-    CommonModule,
-    FormsModule
+    CommonModule
   ],
   exports: [NpFileUploadComponent]
 })