+++ projects/np-ui-lib/src/lib/np-number-box/np-number-box.module.ts
@@ -1,13 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NpNumberBoxComponent } from './np-number-box.component';
-import { FormsModule } from '@angular/forms';
 
 @NgModule({
   declarations: [NpNumberBoxComponent],
   imports: [
-    CommonModule,
-    FormsModule
+    CommonModule
   ],
   exports: [NpNumberBoxComponent]
 })