+++ projects/np-ui-lib/src/lib/np-slider/np-slider.module.ts
@@ -1,13 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NpSliderComponent } from './np-slider.component';
-import { FormsModule } from '@angular/forms';
 
 @NgModule({
   declarations: [NpSliderComponent],
   imports: [
-    CommonModule,
-    FormsModule
+    CommonModule
   ],
   exports: [NpSliderComponent]
 })